+++ CHANGELOG.md
@@ -1,16 +1,20 @@
 ## Unreleased
 
+- Changed: `reflect` is now either a function with signature `(reflecting, set)` or an object with interface `{ withOld(reflecting, old, set) }`. Async-ness is determined by whether the function accepts a `set` parameter, just as for the `derive` callback.
 - Changed: The `reflect` callback now runs *after* subscriptions. Additional `set`/`update` calls while running subscriptions result in `reflect` being called only once with the most recent value.
 - Changed: Changes to which operations cause a subscription-less `writableDerived` to subscribe-then-unsubscribe to all origins:
 	- `set` method calls now do so
 	- Getting old origin values in `reflect` no longer does so
+- Optimization: Origin values are no longer kept if `reflect` will not need them
 - Optimization: Participates in Svelte v3.5.0's [diamond dependencies solution](https://github.com/sveltejs/svelte/pull/2955) (downstream only; there is no diamond dependency handling for reflecting values back to their sources)
-- Optizimation: No more internal writable store; the only internal store is a `derived`
+- Optimization: No more internal writable store; the `derived` is the only internal store remaining
+
+If this update was useful for you, show your appreciation with [money](README.md#--with-money), [kind words](README.md#--with-kind-words), or [a job](README.md#--with-a-job)!
 
 ## 1.0.1 (June 7, 2019)
 
 - Fixed: First subscription was getting its first call before `derive` ran
-- Optimization: Internal derived store is no longer set; all sets go to the internal writable store
+- Optimization: Internal derived store is no longer set; all sets go to the internal writable store (replaced with a different optimization in 2.0.0)
 
 ## 1.0.0 (June 2, 2019)